% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, }{a numeric number.}

\item{b, }{a numeric number.}
}
\value{
The GCD of \code{a} and \code{b}.
}
\description{
Find the GCD of \code{a} and \code{b}.
}
\examples{
euclidean(123612, 13892347912)    
euclidean(100, 1000)
  
}
