% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{first points}
  \item{v2}{second points}
  \item{w}{distance between points}
}}
\usage{
wiki_graph
}
\description{
A test data frame where distance for all edges between 6 nodes are available.
}
\keyword{datasets}
